+++ package.json
@@ -46,6 +46,7 @@
     "ts-loader": "^6.2.2"
   },
   "dependencies": {
+    "@types/chroma-js": "^2.0.0",
     "@types/express": "^4.17.4",
     "@types/morgan": "^1.9.0",
     "@types/webpack": "^4.41.10",
