+++ server/templates/base.html
@@ -138,7 +138,13 @@ endset -%}
       src="https://cdn.jsdelivr.net/npm/cookie-bar/cookiebar-latest.min.js?tracking=1&top=1"
     ></script>
 
-    <script src="https://cdn.plot.ly/plotly-mapbox-1.53.0.min.js"></script>
+    <script 
+      src="https://cdnjs.cloudflare.com/ajax/libs/chroma-js/2.1.0/chroma.min.js" 
+      integrity="sha256-gX8uuyxN8stSMHWO9arnnKyenTOALaVAcVB3b6P87e4=" 
+      crossorigin="anonymous"
+    ></script>
+
+    <script src="https://cdn.plot.ly/plotly-1.53.0.min.js"></script>
     <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/plotly.js/1.51.1/plotly-basic.min.js" integrity="sha256-uFsGuCb3OlYSUV+cZuYWVrTNA+Q5IccRLWCG1vzsCCs=" crossorigin="anonymous"></script> -->
 
     <script
