+++ server/templates/measures.html
@@ -2,101 +2,48 @@
 "base.html" %} {% set title = 'Countermeasures' %} {%- block additional_js_pre
 -%}
 <style type="text/css">
-  @media only screen and (min-width: 992px) and (max-width: 1199px) {
-    .legend {
-      min-height: 473px;
-    }
-  }
-
-  @media only screen and (min-width: 1200px) {
-    .legend {
-      min-height: 555px;
-    }
-  }
-
-  /* apply column-specific scrolling only if columns are side-by-side */
-  /* @media only screen and (min-width: 576px) {
-  html {
-    overflow: hidden;
-  }
-  .container-wrapper {
-    height: 100vh;
-  }
-  .container-left, .container-right {
-    height: inherit;
-    overflow: scroll;
-    padding-bottom: 10em;
-  }
-} */
 </style>
 {% endblock %} {% block content %}
-<!-- Pro bono consulting alert-->
-<!-- <div class="alert alert-warning alert-dismissible fade show" role="alert">
-  <p class="text-center">
-    We're offering pro bono consulting services and custom forecasts for decision-makers. Please reach out <a
-      href="http://epidemicforecasting.org/request-calculation" class="alert-link">here</a>. </p>
-  <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-    <span aria-hidden="true">&times;</span>
-  </button>
-</div> -->
 
-<div class="top-row">
-  <!-- Main model graph -->
-  <div class="main-wrapper">
-    <div class="row">
-      <div class="central-column">
-        <div class="top-row">
-          {{ region_dropdown("regionDropdown") }}
-          <div class="active-infections-block">
-            <span class="number-subheader" id="infections-date"></span>
-            <span class="active-infections"
-              >Active Infections:
-              <span class="infections-estimated" id="infections-estimated"
-                >&mdash;</span
-              >
-              <!-- TODO: Insert official case count into tooltip -->
-              <a href="#case-count-explanation">
-                <span class="question-tooltip">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    height="18"
-                    viewBox="0 0 24 24"
-                    width="18"
-                  >
-                    <path d="M0 0h24v24H0z" fill="none" />
-                    <path
-                      d="M11 18h2v-2h-2v2zm1-16C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8zm0-14c-2.21 0-4 1.79-4 4h2c0-1.1.9-2 2-2s2 .9 2 2c0 2-3 1.75-3 5h2c0-2.25 3-2.5 3-5 0-2.21-1.79-4-4-4z"
-                    />
-                  </svg>
-                </span>
-              </a>
-            </span>
-            <span class="infections-confirmed">
-              Confirmed Infections:
-              <span id="infections-confirmed">&mdash;</span></span
-            >
-          </div>
-
-          <div class="population-block">
-            <span class="number-subheader">Population</span>
-            <span class="infections-population" id="infections-population"
+<!-- Main model graph -->
+<div class="main-wrapper">
+  <div class="row">
+    <div class="central-column">
+      <div class="top-row">
+        {{ region_dropdown("regionDropdown") }}
+        <div class="active-infections-block">
+          <span class="number-subheader" id="infections-date"></span>
+          <span class="active-infections"
+            >Active Infections:
+            <span class="infections-estimated" id="infections-estimated"
               >&mdash;</span
             >
-          </div>
+          </span>
+          <span class="infections-confirmed">
+            Confirmed Infections:
+            <span id="infections-confirmed">&mdash;</span></span
+          >
         </div>
 
-        <hr />
-
-        <div id="current_viz" />
-
-        <hr />
-
-        <div class="model-explanation">
-          <h5 class="mitigation-strength-heading">Explain measures (TODO)</h5>
-          <p>The impact on <span class="selected-region"></span></p>
+        <div class="population-block">
+          <span class="number-subheader">Population</span>
+          <span class="infections-population" id="infections-population"
+            >&mdash;</span
+          >
         </div>
+      </div>
+
+      <hr />
+
+      <div id="current_viz"></div>
+
+      <hr />
+
+      <div class="model-explanation">
+        <h5 class="mitigation-strength-heading">Explain measures (TODO)</h5>
+        <p>The impact on <span class="selected-region"></span></p>
       </div>
     </div>
   </div>
-  {% endblock %} {% block additional_js %} {% endblock %}
 </div>
+{% endblock %} {% block additional_js %} {% endblock %}
